+++ src/components/questionary/questionaryComponents/Boolean/BooleanSearchCriteriaInput.tsx
@@ -1,9 +1,7 @@
 import { FormControl, InputLabel, MenuItem, Select } from '@material-ui/core';
-import React from 'react';
-
+import { SearchCriteriaInputProps } from 'components/common/proposalFilters/QuestionaryFilter';
 import { QuestionFilterCompareOperator } from 'generated/sdk';
-
-import { SearchCriteriaInputProps } from '../QuestionaryFilter';
+import React from 'react';
 
 function BooleanSearchCriteriaInput({ onChange }: SearchCriteriaInputProps) {
   return (
