+++ src/components/questionary/questionaryComponents/DatePicker/DateSearchCriteriaInput.tsx
@@ -11,11 +11,9 @@ import {
   KeyboardDatePicker,
   MuiPickersUtilsProvider,
 } from '@material-ui/pickers';
-import React, { useState } from 'react';
-
+import { SearchCriteriaInputProps } from 'components/common/proposalFilters/QuestionaryFilter';
 import { QuestionFilterCompareOperator } from 'generated/sdk';
-
-import { SearchCriteriaInputProps } from '../QuestionaryFilter';
+import React, { useState } from 'react';
 
 function DateSearchCriteriaInput({ onChange }: SearchCriteriaInputProps) {
   const [value, setValue] = useState<string | null>(null);
