+++ src/components/questionary/questionaryComponents/MultipleChoice/MultipleChoiceDefinition.tsx
@@ -2,6 +2,7 @@ import RadioButtonCheckedIcon from '@material-ui/icons/RadioButtonChecked';
 import React from 'react';
 
 import defaultRenderer from 'components/questionary/DefaultQuestionRenderer';
+import MultipleChoiceSearchCriteriaComponent from 'components/questionary/questionaryComponents/MultipleChoice/MultipleChoiceSearchCriteriaComponent';
 import { DataType } from 'generated/sdk';
 
 import { QuestionaryComponentDefinition } from '../../QuestionaryComponentRegistry';
@@ -26,4 +27,5 @@ export const multipleChoiceDefinition: QuestionaryComponentDefinition = {
   },
   createYupValidationSchema: createMultipleChoiceValidationSchema,
   getYupInitialValue: ({ answer }) => answer.value || [],
+  searchCriteriaComponent: MultipleChoiceSearchCriteriaComponent,
 };
