+++ src/components/questionary/questionaryComponents/MultipleChoice/MultipleChoiceSearchCriteriaComponent.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { SearchCriteriaInputProps } from '../../../common/proposalFilters/QuestionaryFilter';
+
+function MultipleChoiceSearchCriteriaComponent({
+  onChange,
+}: SearchCriteriaInputProps) {
+  return <div></div>;
+}
+
+export default MultipleChoiceSearchCriteriaComponent;
