+++ src/components/questionary/questionaryComponents/TextInput/TextSearchCriteriaInput.tsx
@@ -6,11 +6,9 @@ import {
   Select,
   TextField,
 } from '@material-ui/core';
-import React, { useState } from 'react';
-
+import { SearchCriteriaInputProps } from 'components/common/proposalFilters/QuestionaryFilter';
 import { QuestionFilterCompareOperator } from 'generated/sdk';
-
-import { SearchCriteriaInputProps } from '../QuestionaryFilter';
+import React, { useState } from 'react';
 
 function TextSearchCriteriaComponent({ onChange }: SearchCriteriaInputProps) {
   const [value, setValue] = useState('');
