+++ src/components/questionary/questionaryComponents/UnknownSearchCriteriaInput.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-
-import { SearchCriteriaInputProps } from '../QuestionaryFilter';
+import { SearchCriteriaInputProps } from '../../common/proposalFilters/QuestionaryFilter';
 
 function UnknownSearchCriteriaInput({ question }: SearchCriteriaInputProps) {
   return <span>{`Search input not implemented for ${question.dataType}`}</span>;
